Move phys to buddy
